move files & optimize for future archs
